+++ package.json
@@ -61,7 +61,6 @@
     "prettier": "2.8.0",
     "react": "^18.2.0",
     "react-aria": "^3.27.0",
-    "react-cookie": "^6.1.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.46.1",
     "rehype-pretty-code": "^0.10.0",
@@ -77,7 +76,6 @@
     "toml": "^3.0.0",
     "unified": "^10.1.2",
     "unist-util-visit": "^5.0.0",
-    "universal-cookie": "^6.1.1",
     "updates": "^15.0.2",
     "vfile": "^6.0.1"
   },
