+++ src/components/ThemeToggler.tsx
@@ -1,36 +1,30 @@
 import { cssObj } from '@fuel-ui/css';
-import { Box, Icon } from '@fuel-ui/react';
+import { Box, Icon, loadIcons } from '@fuel-ui/react';
 import type { ElementRef } from 'react';
 import { useEffect, useRef } from 'react';
-import { useCookies } from 'react-cookie';
 
 import useTheme from '../hooks/useTheme';
 
+loadIcons('/icons/sprite.svg');
+
 export default function ThemeToggler() {
-  const { theme: current, setTheme: setFuelTheme } = useTheme();
+  const { theme, setTheme: setFuelTheme } = useTheme();
   const ref = useRef<ElementRef<'div'>>(null);
-  const [, setCookie] = useCookies(['theme']);
 
   const handleChange = async () => {
-    const next = current === 'dark' ? 'light' : 'dark';
+    const next = theme === 'dark' ? 'light' : 'dark';
     setFuelTheme(next);
-    setCookie('theme', next, { path: '/' });
     document.documentElement.setAttribute('data-theme', next);
     ref.current?.setAttribute('data-theme', next);
   };
 
   useEffect(() => {
-    document.documentElement.setAttribute('data-theme', current);
-    ref.current?.setAttribute('data-theme', current);
-  }, []);
+    document.documentElement.setAttribute('data-theme', theme);
+    ref.current?.setAttribute('data-theme', theme);
+  }, [theme]);
 
   return (
-    <Box
-      ref={ref}
-      data-theme={current}
-      css={styles.root}
-      onClick={handleChange}
-    >
+    <Box ref={ref} data-theme={theme} css={styles.root} onClick={handleChange}>
       <Icon icon="SunFilled" size={20} stroke={1.2} />
       <Icon icon="MoonStars" size={20} stroke={1.2} />
     </Box>
